move firebase request off the cli process
